Added plugins functionality
